pyproject.toml
[project]
name = "ims-telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4==4.12.3",
    "bs4==0.0.2",
    "httpx==0.27.0",
    "mysql-connector-python==9.0.0",
    "pytelegrambotapi==4.20.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.0.1",
    "python-telegram-bot==21.3",
    "requests==2.32.3",
    "telebot==0.0.5",
]
